auth-service/requirements.txt
Flask==3.0.3
Flask-Cors==4.0.0
PyJWT==2.8.0
passlib==1.7.4
Flask-SQLAlchemy==3.0.3
psycopg2-binary==2.9.10
